DistanceEuclidienne

+++ src/main/java/fr/groupeh6/sae/distance/DistanceEuclidienne.java
@@ -9,9 +9,9 @@ public class DistanceEuclidienne implements Distance {
 
 	@Override
 	public double distance(IPoint p1, IPoint p2, List<Column> columns) {
-		int somme = 0;
+		double somme = 0;
 		for(Column column : columns) {
-			somme +=Math.pow(1.0*p1.getNormalizedValue(column)-p2.getNormalizedValue(column), 2);
+			somme += Math.pow(1.0*p1.getNormalizedValue(column)-p2.getNormalizedValue(column), 2);
 		}
 		return Math.sqrt(somme);
 	}
